requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
flask-restful
flasgger

opencv-python==4.10.0.84